+++ internal/application/user_service.go
@@ -9,7 +9,7 @@ import (
 type UserSerVice interface {
 	RegisterUser(c context.Context, req model.SignUpRequest) (bool, error)
 	Login(c context.Context, req model.LoginRequest) (model.LoginResponse, error)
-	GetUserInfo(c context.Context, userID int) (*model.UserInfoResponse, error)
+	GetUserInfo(c context.Context, userID int64) (*model.UserInfoResponse, error)
 	UpdateUserInfo(c context.Context, req model.UserUpdateRequest) (bool, error)
 	ChangePassword(c context.Context, req model.UserUpdatePassRequest) (bool, error)
 	Follow(c context.Context, req model.FollowRequest) (bool, error)
